GRPC/backend/requirements.txt
grpcio 
grpcio-tools 
googleapis-common-protos
fastapi
uvicorn